Move ScrollFlagArea paint event from CodeEditor to ScrollFlagArea class.
